requirements.txt
click
flask
jsonschema
mysqlclient
pyyaml
